backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.0
scikit-learn==1.5.2
numpy==2.1.3
pytest==8.3.4
httpx==0.28.1
ruff==0.8.3
